+++ packages/dapp-components/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @cfx-kit/dapp-components
 
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @cfx-kit/dapp-utils@0.0.23
+
+## 0.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @cfx-kit/dapp-utils@0.0.22
+
+# @cfx-kit/dapp-components
+
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies
+  - @cfx-kit/dapp-utils@0.0.21
+
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @cfx-kit/dapp-utils@0.0.20
+
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - @cfx-kit/dapp-utils@0.0.19
+
 ## 0.0.18
 
 ### Patch Changes